docs: updated SCRAM-hashed password options in docs, chore: moved saslprep to separate module
